services/trade-execution-engine/requirements.txt
# Trade Execution Engine Requirements
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
openai==1.3.7
cryptography>=41.0.0
PyJWT==2.8.0
python-multipart==0.0.6

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Additional dependencies
pandas>=1.5.0
numpy>=1.21.0
